refactor(workbench): wire layout and extracted domain hooks
Move pane open/close/resize into useWorkbenchLayout and switch
AppWorkbench onto the search, export, and sandbox/reliability hooks
so those domains own their state. Combined orchestration size drops;
gate ceilings ratchet in the follow-up commit.
